WIP: clean up walberla impl
